fine-tuning/pyproject.toml
[project]
name = "docintel-training"
version = "0.1.0"
description = "Fine-tuning pipeline for DocIntel document intelligence models"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "unsloth>=2024.8",
    "transformers>=4.44.0",
    "trl>=0.9.0",
    "peft>=0.12.0",
    "accelerate>=0.33.0",
    "bitsandbytes>=0.43.0",
    "optimum>=1.21.0",
    "onnx>=1.16.0",
    "onnxruntime>=1.18.0",
    "olive-ai>=0.7.0",
    "wandb>=0.17.0",
    "pyyaml>=6.0",
    "datasets>=2.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
